scraper/requirements.txt
numpy==1.23.5
pandas==1.5.3
requests==2.31.0
beautifulsoup4==4.12.2
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
schedule==1.2.1
urllib3>=2.0.0
lxml==4.9.3
fake-useragent==0.1.11
selenium==4.15.0


